# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Marketing Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/{businessId}/content:
    get:
      operationId: fetchContent
      tags: [content]
      summary: List generated content for a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of content pieces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentPiece"
    post:
      operationId: generateContent
      tags: [content]
      summary: Generate AI content (streams SSE)
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentRequest"
      responses:
        "200":
          description: SSE stream of generated content
          content:
            text/event-stream: {}
  /businesses/{businessId}/content/{id}:
    get:
      operationId: getContent
      tags: [content]
      summary: Get a content piece
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content piece
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentPiece"
    delete:
      operationId: deleteContent
      tags: [content]
      summary: Delete a content piece
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /businesses/{businessId}/content/recent:
    get:
      operationId: listRecentContent
      tags: [content]
      summary: Get recent content across all types for a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recent content pieces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentPiece"

  # --- Campaigns ---
...
  /businesses/{businessId}/competitors:
    get:
      operationId: listCompetitors
      tags: [competitors]
      summary: List competitors for a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of competitors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Competitor"
    post:
      operationId: createCompetitor
      tags: [competitors]
      summary: Add a competitor
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitorInput"
      responses:
        "201":
          description: Created competitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competitor"
  /businesses/{businessId}/competitors/{id}:
    delete:
      operationId: deleteCompetitor
      tags: [competitors]
      summary: Remove a competitor
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /businesses/{businessId}/competitors/{id}/analyze:
    post:
      operationId: analyzeCompetitor
      tags: [competitors]
      summary: Run AI competitor analysis (streams SSE)
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SSE stream of competitor analysis
          content:
            text/event-stream: {}

  # --- OpenAI Chat (marketing assistant) ---
...
components:
  schemas:
...
    ContentPiece:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        type:
          type: string
          enum: [blog, social, email, ad, strategy]
        title:
          type: string
        body:
          type: string
        platform:
          type: ["string", "null"]
        tone:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published, archived]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - businessId
        - type
        - title
        - body
        - status
        - createdAt
    ContentRequest:
      type: object
      properties:
        type:
          type: string
          enum: [blog, social, email, ad, strategy]
        topic:
          type: string
        tone:
          type: string
        platform:
          type: string
        additionalContext:
          type: string
      required:
        - type
        - topic

    # --- Campaign schemas ---
...
    Competitor:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        name:
          type: string
        website:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        strengths:
          type: ["string", "null"]
        weaknesses:
          type: ["string", "null"]
        lastAnalyzedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - businessId
        - name
        - createdAt
    CompetitorInput:
      type: object
      properties:
        name:
          type: string
        website:
          type: string
        notes:
          type: string
      required:
        - name

    # --- OpenAI schemas ---
...