# lib/api-zod/src/generated/types/competitor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Marketing Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Competitor {
  id: number;
  businessId: number;
  name: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  strengths?: string | null;
  /** @nullable */
  weaknesses?: string | null;
  /** @nullable */
  lastAnalyzedAt?: Date | null;
  createdAt: Date;
}