# lib/api-zod/src/generated/types/competitorInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Marketing Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface CompetitorInput {
  name: string;
  website?: string;
  notes?: string;
}